requirements.txt
openai>=1.45.0
python-dotenv>=1.0.1
PyGithub>=2.3.0
GitPython>=3.1.43
PyYAML>=6.0.2
tenacity>=8.3.0
rich>=13.7.1
pytest>=8.3.3
requests>=2.32.3
jsonschema>=4.23.0
numpy>=1.26.0
sounddevice>=0.4.6
pdfplumber>=0.10.4
python-docx>=1.1.2
scikit-learn>=1.4.2
faster-whisper>=1.0.3
